Add an option for the map pane in which to draw clusters.
The default is 'overlayMouseTarget', which can respond to DOM events.
If you don't need to respond to DOM events (e.g. if clickToZoom is
disabled), then you can choose a different layer, such as if you want
to have other markers drawn above the cluster markers instead of
beneath them.